initial implemenation of schema database manager
Added a python script which simplifies the task to dump and load
database dumps. No more need for the multistep process to
get into PHPMyAdmin via our hosting provider.

Currently can dump and load "baseline data" which right now is
users and activity types. I am planning to also support dumping
and loading of the table structure as well, as non-destructively as
possible.